ai-city-service/requirements.txt
fastapi
uvicorn
celery
pika
python-dotenv
supabase
google-generativeai
requests
python-multipart     # (only needed if FastAPI ever handles file uploads)
